backend/requirements.txt
django==4.0
djangorestframework==3.13.1
PyJWT==2.4.0
django-filter==22.1
gunicorn==20.0.4
psycopg2-binary==2.9.3
asgiref==3.5.2
djangorestframework-simplejwt==4.8.0
sqlparse==0.4.2
python-dotenv==0.20.0
pillow==9.2.0
drf-extra-fields==3.4.0
djoser==2.1.0